LayoutTests/imported/w3c/web-platform-tests/cssom/overflow-serialization-expected.txt


FAIL CSSOM - Overlow property has different serialization than other shorthands. assert_equals: Overflow-x/y longhands with same CSS-wide keyword should serialize correctly. expected "overflow: initial;" but got ""
